requirements.txt
Django>=6.0.4
djangorestframework>=3.15.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
gunicorn>=21.2.0
